Change the hashing algorithm
Reverse the hashing algorithm so that different days have substantially
different hash codes. Previously, they'd only differ by, like, one. By
flipping the game ID first, we make it so that days a thousand years
apart will get neighboring hash codes -- something I don't really care
about.

This also changes the way that the data is keyed in the localStorage.
Not only does it introduce the language (in anticipation of multiple
language support) but it also keys not by game ID but by the letters.
This means that if the same puzzle ever comes up again, we'll see it
already solved.