Regenerate the date at runtime
When the default game ID is requested, if one isn't passed in, then
"today" is used. Instead of storing this in a constant with app-long
lifetime, regenerate it at runtime. This means that after midnight,
clicking the header will fast-forward the user to the new day (which is
to be expected).